chore(data-table): remove storybook from next
