+++ packages/react/src/components/DataTable/DataTable.mdx
@@ -195,7 +195,7 @@ The `DataTable` components supports row-level expansion when combined with the
 `TableExpandHeader`, `TableExpandRow`, and `TableExpandedRow` components.
 
 <Preview>
-  <Story id="components-datatable-expansion--usage" />
+  <Story id="components-datatable-expansion--playground" />
 </Preview>
 
 _Note: press "Show code" above to view a code snippet of this example_
@@ -285,7 +285,7 @@ _Note: press "Show code" above to view a code snippet of this example_
 ## Toolbar
 
 <Preview>
-  <Story id="components-datatable-toolbar--default" />
+  <Story id="components-datatable-toolbar--playground" />
 </Preview>
 
 ### Overflow Menu
