--- packages/react/src/components/DataTable/next/dynamic-content/story.scss
@@ -1,3 +1,0 @@
-.demo-expanded-td td {
-  padding-left: 4rem;
-}
