--- packages/react/src/components/DataTable/next/expansion/DataTable-expansion-story.scss
@@ -1,12 +1,0 @@
-@use '../../../../../scss/type';
-@use '../../../../../scss/theme';
-
-.demo-inner-container-header {
-  @include type.type-style('heading-compact-01');
-  color: theme.$text-primary;
-}
-
-.demo-inner-container-content {
-  @include type.type-style('body-01');
-  color: theme.$text-secondary;
-}
