--- packages/react/src/components/DataTable/stories/DataTable-toolbar-story.js
@@ -1,271 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2020
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import { action } from '@storybook/addon-actions';
-import React from 'react';
-import Button from '../../Button';
-import OverflowMenu from '../../OverflowMenu';
-import OverflowMenuItem from '../../OverflowMenuItem';
-import {
-  default as DataTable,
-  TableContainer,
-  Table,
-  TableHead,
-  TableRow,
-  TableHeader,
-  TableBody,
-  TableCell,
-  TableToolbar,
-  TableToolbarAction,
-  TableToolbarContent,
-  TableToolbarSearch,
-  TableToolbarMenu,
-} from '../../DataTable';
-import mdx from '../DataTable.mdx';
-import { headers, rows } from './shared';
-
-export default {
-  title: 'Components/DataTable/Toolbar',
-  component: DataTable,
-  subcomponents: {
-    TableContainer,
-    Table,
-    TableHead,
-    TableRow,
-    TableHeader,
-    TableBody,
-    TableCell,
-  },
-  parameters: {
-    docs: {
-      page: mdx,
-    },
-  },
-};
-
-export const DefaultToolbar = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      getToolbarProps,
-      onInputChange,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With toolbar"
-        {...getTableContainerProps()}>
-        <TableToolbar {...getToolbarProps()} aria-label="data table toolbar">
-          <TableToolbarContent>
-            <TableToolbarSearch
-              onChange={onInputChange}
-              onClear={action('onClear')}
-              onFocus={(event, handleExpand) => {
-                action('onFocus')();
-                handleExpand(event, true);
-              }}
-              onBlur={(event, handleExpand) => {
-                action('onBlur')();
-                const { value } = event.target;
-                if (!value) {
-                  handleExpand(event, false);
-                }
-              }}
-            />
-            <TableToolbarMenu light>
-              <TableToolbarAction onClick={action('Action 1 Click')}>
-                Action 1
-              </TableToolbarAction>
-              <TableToolbarAction onClick={action('Action 2 Click')}>
-                Action 2
-              </TableToolbarAction>
-              <TableToolbarAction onClick={action('Action 3 Click')}>
-                Action 3
-              </TableToolbarAction>
-            </TableToolbarMenu>
-            <Button onClick={action('Button click')}>Primary Button</Button>
-          </TableToolbarContent>
-        </TableToolbar>
-        <Table {...getTableProps()}>
-          <TableHead>
-            <TableRow>
-              {headers.map((header) => (
-                <TableHeader key={header.key} {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <TableRow key={row.id} {...getRowProps({ row })}>
-                {row.cells.map((cell) => (
-                  <TableCell key={cell.id}>{cell.value}</TableCell>
-                ))}
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
-
-export const SmallToolbar = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      getToolbarProps,
-      onInputChange,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With toolbar"
-        {...getTableContainerProps()}>
-        <TableToolbar
-          {...getToolbarProps()}
-          aria-label="data table toolbar"
-          size="sm">
-          <TableToolbarContent>
-            <TableToolbarSearch
-              onChange={onInputChange}
-              onClear={action('onClear')}
-              onFocus={(event, handleExpand) => {
-                action('onFocus')();
-                handleExpand(event, true);
-              }}
-              onBlur={(event, handleExpand) => {
-                action('onBlur')();
-                const { value } = event.target;
-                if (!value) {
-                  handleExpand(event, false);
-                }
-              }}
-            />
-            <TableToolbarMenu light size="sm">
-              <TableToolbarAction onClick={action('Action 1 Click')}>
-                Action 1
-              </TableToolbarAction>
-              <TableToolbarAction onClick={action('Action 2 Click')}>
-                Action 2
-              </TableToolbarAction>
-              <TableToolbarAction onClick={action('Action 3 Click')}>
-                Action 3
-              </TableToolbarAction>
-            </TableToolbarMenu>
-            <Button onClick={action('Button click')}>Primary Button</Button>
-          </TableToolbarContent>
-        </TableToolbar>
-        <Table {...getTableProps()}>
-          <TableHead>
-            <TableRow>
-              {headers.map((header) => (
-                <TableHeader key={header.key} {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <TableRow key={row.id} {...getRowProps({ row })}>
-                {row.cells.map((cell) => (
-                  <TableCell key={cell.id}>{cell.value}</TableCell>
-                ))}
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
-
-export const WithOverflowMenu = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      getToolbarProps,
-      onInputChange,
-    }) => (
-      <TableContainer title="DataTable" description="With overflow menu">
-        <TableToolbar {...getToolbarProps()} aria-label="data table toolbar">
-          <TableToolbarContent>
-            <TableToolbarSearch
-              onChange={onInputChange}
-              onClear={action('onClear')}
-              onFocus={(event, handleExpand) => {
-                action('onFocus')();
-                handleExpand(event, true);
-              }}
-              onBlur={(event, handleExpand) => {
-                action('onBlur')();
-                const { value } = event.target;
-                if (!value) {
-                  handleExpand(event, false);
-                }
-              }}
-            />
-            <TableToolbarMenu light>
-              <TableToolbarAction onClick={action('Action 1 Click')}>
-                Action 1
-              </TableToolbarAction>
-              <TableToolbarAction onClick={action('Action 2 Click')}>
-                Action 2
-              </TableToolbarAction>
-              <TableToolbarAction onClick={action('Action 3 Click')}>
-                Action 3
-              </TableToolbarAction>
-            </TableToolbarMenu>
-            <Button onClick={action('Button click')}>Primary Button</Button>
-          </TableToolbarContent>
-        </TableToolbar>
-        <Table {...getTableProps()}>
-          <TableHead>
-            <TableRow>
-              {headers.map((header) => (
-                <TableHeader key={header.key} {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-              <TableHeader />
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <TableRow key={row.id} {...getRowProps({ row })}>
-                {row.cells.map((cell) => (
-                  <TableCell key={cell.id}>{cell.value}</TableCell>
-                ))}
-                <TableCell className="cds--table-column-menu">
-                  <OverflowMenu size="sm" light flipped>
-                    <OverflowMenuItem>Action 1</OverflowMenuItem>
-                    <OverflowMenuItem>Action 2</OverflowMenuItem>
-                    <OverflowMenuItem>Action 3</OverflowMenuItem>
-                  </OverflowMenu>
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
