--- packages/react/src/components/DataTable/stories/expansion/DataTable-expansion-story.js
@@ -1,387 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2018
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import './DataTable-expansion-story.scss';
-import { action } from '@storybook/addon-actions';
-
-import React from 'react';
-import Button from '../../../Button';
-import DataTable, {
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableExpandHeader,
-  TableExpandRow,
-  TableExpandedRow,
-  TableHead,
-  TableHeader,
-  TableRow,
-  TableToolbar,
-  TableToolbarAction,
-  TableToolbarContent,
-  TableToolbarSearch,
-  TableToolbarMenu,
-} from '../../../DataTable';
-import { rows, headers } from '../shared';
-import mdx from '../../DataTable.mdx';
-
-export default {
-  title: 'Components/DataTable/Expansion',
-  component: DataTable,
-  subcomponents: {
-    TableExpandHeader,
-    TableExpandRow,
-    TableExpandedRow,
-    Table,
-    TableBody,
-    TableCell,
-    TableContainer,
-    TableHead,
-    TableHeader,
-    TableRow,
-  },
-  parameters: {
-    docs: {
-      page: mdx,
-    },
-  },
-};
-
-export const Usage = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With expansion"
-        {...getTableContainerProps()}>
-        <Table {...getTableProps()}>
-          <TableHead>
-            <TableRow>
-              <TableExpandHeader id="expand" />
-              {headers.map((header, i) => (
-                <TableHeader
-                  id={header.header}
-                  key={i}
-                  {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <React.Fragment key={row.id}>
-                <TableExpandRow {...getRowProps({ row })}>
-                  {row.cells.map((cell) => (
-                    <TableCell key={cell.id}>{cell.value}</TableCell>
-                  ))}
-                </TableExpandRow>
-                <TableExpandedRow
-                  colSpan={headers.length + 1}
-                  className="demo-expanded-td">
-                  <h6>Expandable row content</h6>
-                  <div>Description here</div>
-                </TableExpandedRow>
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
-
-export const ExtraSmallExpansion = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      onInputChange,
-      getToolbarProps,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With expansion"
-        {...getTableContainerProps()}>
-        <TableToolbar {...getToolbarProps()} size="xs">
-          <TableToolbarContent>
-            <TableToolbarSearch persistent="true" onChange={onInputChange} />
-            <TableToolbarMenu>
-              <TableToolbarAction onClick={() => alert('Alert 1')}>
-                Action 1
-              </TableToolbarAction>
-              <TableToolbarAction onClick={() => alert('Alert 2')}>
-                Action 2
-              </TableToolbarAction>
-              <TableToolbarAction onClick={() => alert('Alert 3')}>
-                Action 3
-              </TableToolbarAction>
-            </TableToolbarMenu>
-            <Button onClick={action('Add new row')} size="small" kind="primary">
-              Add new
-            </Button>
-          </TableToolbarContent>
-        </TableToolbar>
-        <Table {...getTableProps()} size="compact">
-          <TableHead>
-            <TableRow>
-              <TableExpandHeader id="expand" />
-              {headers.map((header, i) => (
-                <TableHeader
-                  id={header.header}
-                  key={i}
-                  {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <React.Fragment key={row.id}>
-                <TableExpandRow {...getRowProps({ row })}>
-                  {row.cells.map((cell) => (
-                    <TableCell key={cell.id}>{cell.value}</TableCell>
-                  ))}
-                </TableExpandRow>
-                <TableExpandedRow
-                  colSpan={headers.length + 1}
-                  className="demo-expanded-td">
-                  <h6>Expandable row content</h6>
-                  <div>Description here</div>
-                </TableExpandedRow>
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
-
-export const SmallExpansion = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With expansion"
-        {...getTableContainerProps()}>
-        <Table {...getTableProps()} size="sm">
-          <TableHead>
-            <TableRow>
-              <TableExpandHeader id="expand" />
-              {headers.map((header, i) => (
-                <TableHeader
-                  id={header.header}
-                  key={i}
-                  {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <React.Fragment key={row.id}>
-                <TableExpandRow {...getRowProps({ row })}>
-                  {row.cells.map((cell) => (
-                    <TableCell key={cell.id}>{cell.value}</TableCell>
-                  ))}
-                </TableExpandRow>
-                <TableExpandedRow
-                  colSpan={headers.length + 1}
-                  className="demo-expanded-td">
-                  <h6>Expandable row content</h6>
-                  <div>Description here</div>
-                </TableExpandedRow>
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
-
-export const MediumExpansion = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With expansion"
-        {...getTableContainerProps()}>
-        <Table {...getTableProps()} size="md">
-          <TableHead>
-            <TableRow>
-              <TableExpandHeader id="expand" />
-              {headers.map((header, i) => (
-                <TableHeader
-                  id={header.header}
-                  key={i}
-                  {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <React.Fragment key={row.id}>
-                <TableExpandRow {...getRowProps({ row })}>
-                  {row.cells.map((cell) => (
-                    <TableCell key={cell.id}>{cell.value}</TableCell>
-                  ))}
-                </TableExpandRow>
-                <TableExpandedRow
-                  colSpan={headers.length + 1}
-                  className="demo-expanded-td">
-                  <h6>Expandable row content</h6>
-                  <div>Description here</div>
-                </TableExpandedRow>
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
-
-export const ExtraLargeExpansion = () => (
-  <DataTable rows={rows} headers={headers}>
-    {({
-      rows,
-      headers,
-      getHeaderProps,
-      getRowProps,
-      getTableProps,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With expansion"
-        {...getTableContainerProps()}>
-        <Table {...getTableProps()} size="xl">
-          <TableHead>
-            <TableRow>
-              <TableExpandHeader id="expand" />
-              {headers.map((header, i) => (
-                <TableHeader
-                  id={header.header}
-                  key={i}
-                  {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <React.Fragment key={row.id}>
-                <TableExpandRow {...getRowProps({ row })}>
-                  {row.cells.map((cell) => (
-                    <TableCell key={cell.id}>{cell.value}</TableCell>
-                  ))}
-                </TableExpandRow>
-                <TableExpandedRow
-                  colSpan={headers.length + 1}
-                  className="demo-expanded-td">
-                  <h6>Expandable row content</h6>
-                  <div>Description here</div>
-                </TableExpandedRow>
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  </DataTable>
-);
-
-export const BatchExpansion = () => (
-  <DataTable
-    rows={rows}
-    headers={headers}
-    render={({
-      rows,
-      headers,
-      getHeaderProps,
-      getExpandHeaderProps,
-      getRowProps,
-      getTableProps,
-      getTableContainerProps,
-    }) => (
-      <TableContainer
-        title="DataTable"
-        description="With batch expansion"
-        {...getTableContainerProps()}>
-        <Table {...getTableProps()}>
-          <TableHead>
-            <TableRow>
-              <TableExpandHeader
-                id="expand"
-                enableToggle
-                {...getExpandHeaderProps()}
-              />
-              {headers.map((header, i) => (
-                <TableHeader
-                  id={header.header}
-                  key={i}
-                  {...getHeaderProps({ header })}>
-                  {header.header}
-                </TableHeader>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows.map((row) => (
-              <React.Fragment key={row.id}>
-                <TableExpandRow {...getRowProps({ row })}>
-                  {row.cells.map((cell) => (
-                    <TableCell key={cell.id}>{cell.value}</TableCell>
-                  ))}
-                </TableExpandRow>
-                <TableExpandedRow
-                  colSpan={headers.length + 1}
-                  className="demo-expanded-td">
-                  <h6>Expandable row content</h6>
-                  <div>Description here</div>
-                </TableExpandedRow>
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    )}
-  />
-);
